[SPARK-7358] Move DataFrame mathfunctions into functions
